Se reparo la carga de imagenes de perfil en la db, se reparo ademas la tabla debil gps_data_user no referenciaba a la columna gps_data, las vcard traen los datos de busco y ofrezco y general el mini busco y el miniofrezco.

+++ backend/src/main/resources/schema.sql
@@ -77,7 +77,7 @@ KEY (`latitude`, `longitude`)
 DROP TABLE IF EXISTS `gps_data_users`;
 		
 CREATE TABLE `gps_data_users` (
-  `id_gps_record` INTEGER NOT NULL,
+  `id_gps_record` BIGINT NOT NULL,
   `id_user` BIGINT NOT NULL,
   PRIMARY KEY (`id_gps_record`, `id_user`)
 );
